rename `utilityFunctions` and `utilityTypes` to just `functions` and `types`
