+++ src/functions/misc.ts
@@ -1,5 +1,5 @@
-import { Stringable, TemplateLiteralStringable, ToString } from '../utilityTypes/String'
-import { Entries, Equals, OnlyInfer } from '../utilityTypes/misc'
+import { Stringable, TemplateLiteralStringable, ToString } from '../types/String'
+import { Entries, Equals, OnlyInfer } from '../types/misc'
 import { isEqual } from 'lodash'
 import { Throw } from 'throw-expression'
 import { Narrow } from 'ts-toolbelt/out/Function/Narrow'
