+++ src/types/String.ts
