+++ test/functions/Function.spec.ts
@@ -1,5 +1,5 @@
-import { bindThis } from '../../src/utilityFunctions/Function'
-import { exactly } from '../../src/utilityFunctions/misc'
+import { bindThis } from '../../src/functions/Function'
+import { exactly } from '../../src/functions/misc'
 import { PowerAssert } from 'typed-nodejs-assert'
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires,@typescript-eslint/no-unsafe-assignment -- https://github.com/detachHead/typed-nodejs-assert#with-power-assert
