+++ test/functions/Number.spec.ts
@@ -14,9 +14,9 @@ import {
     rightShift,
     subtract,
     toNumber,
-} from '../../src/utilityFunctions/Number'
-import { exactly } from '../../src/utilityFunctions/misc'
-import { TupleOf } from '../../src/utilityTypes/Array'
+} from '../../src/functions/Number'
+import { exactly } from '../../src/functions/misc'
+import { TupleOf } from '../../src/types/Array'
 import { PowerAssert } from 'typed-nodejs-assert'
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires,@typescript-eslint/no-unsafe-assignment -- https://github.com/detachHead/typed-nodejs-assert#with-power-assert
